Add ConditionalCaseBase to reduce duplication
